Premezal som celu Databazu lebo som musel menit udaje ADMINA nemusis pridavat noveho, prihlasenie do admina funguje aj bey toho aby bol v databaze
Zmenil som datatyp z KG a DLZKY na Double / v SQL REAL
odstranil cislo_reviru to je to iste ako IDreviru

da sa pridavat ulovok a pridava sa do DB

+++ src/main/java/org/projekt/Ulovok.java
@@ -16,11 +16,11 @@ public class Ulovok {
     private int cisloReviru;
     private String druhRyby;
     //private int pocetKusov;
-    private int dlzkaVcm;
-    private int hmotnostVkg;
+    private double dlzkaVcm;
+    private double hmotnostVkg;
     private int kontrola;
 
-    public Ulovok(LocalDate datumUlovku, int cisloReviru, String druhRyby, int dlzkaVcm, int hmotnostVkg, int kontrola) {
+    public Ulovok(LocalDate datumUlovku, int cisloReviru, String druhRyby, double dlzkaVcm, double hmotnostVkg, int kontrola) {
         this.idUlovok = idUlovok;
         this.datumUlovku = datumUlovku;
         this.cisloReviru = cisloReviru;
